Improved internal organisation
