+++ lib/src/layer/polyline_layer/polyline_layer.dart
@@ -10,6 +10,7 @@ import 'package:flutter_map/src/misc/simplify.dart';
 import 'package:latlong2/latlong.dart';
 
 part 'painter.dart';
+part 'pattern.dart';
 part 'polyline.dart';
 part 'projected_polyline.dart';
